+++ blueprints/flexberry-group/index.ts
@@ -75,6 +75,9 @@ class GroupBlueprint {
       case 'flexberry-serializer-init':
         this.emberGenerate("models", true, "app/serializers");
         break;
+      default:
+        throw new Error(`Unknown blueprint: ${this.blueprintName}`);
+
     }
   }